fixed images to show in listing
