+++ partFinder/HomeViewModel.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import SwiftUI
+import Combine
 
 class HomeViewModel: ObservableObject {
     @Published var selectedCity: String = "Location"
